refactor(tasks): change name of sequence tagging to sequence labelling
