+++ embeddings/pipeline/hugging_face_POS_tagging.py
@@ -8,10 +8,10 @@ from embeddings.data.hugging_face_data_loader import HuggingFaceDataLoader
 from embeddings.data.hugging_face_dataset import HuggingFaceDataset
 from embeddings.data.io import T_path
 from embeddings.embedding.flair_embedding import FlairTransformerWordEmbedding
-from embeddings.evaluator.sequence_tagging_evaluator import POSTaggingEvaluator
+from embeddings.evaluator.sequence_labeling_evaluator import POSTaggingEvaluator
 from embeddings.model.flair_model import FlairModel
 from embeddings.pipeline.standard_pipeline import StandardPipeline
-from embeddings.task.flair_task.sequence_tagging import SequenceTagging
+from embeddings.task.flair_task.sequence_labeling import SequenceLabeling
 from embeddings.transformation.flair_transformation.column_corpus_transformation import (
     ColumnCorpusTransformation,
 )
@@ -33,7 +33,7 @@ class HuggingFacePOSTaggingPipeline(
         data_loader = HuggingFaceDataLoader()
         transformation = ColumnCorpusTransformation(input_column_name, target_column_name)
         embedding = FlairTransformerWordEmbedding(embedding_name)
-        task = SequenceTagging(output_path, hidden_size=hidden_size)
+        task = SequenceLabeling(output_path, hidden_size=hidden_size)
         model = FlairModel(embedding, task)
         evaluator = POSTaggingEvaluator()
         super().__init__(dataset, data_loader, transformation, model, evaluator)
