+++ embeddings/pipeline/hugging_face_sequence_labeling.py
@@ -8,16 +8,16 @@ from embeddings.data.hugging_face_data_loader import HuggingFaceDataLoader
 from embeddings.data.hugging_face_dataset import HuggingFaceDataset
 from embeddings.data.io import T_path
 from embeddings.embedding.flair_embedding import FlairTransformerWordEmbedding
-from embeddings.evaluator.sequence_tagging_evaluator import SequenceTaggingEvaluator
+from embeddings.evaluator.sequence_labeling_evaluator import SequenceLabelingEvaluator
 from embeddings.model.flair_model import FlairModel
 from embeddings.pipeline.standard_pipeline import StandardPipeline
-from embeddings.task.flair_task.sequence_tagging import SequenceTagging
+from embeddings.task.flair_task.sequence_labeling import SequenceLabeling
 from embeddings.transformation.flair_transformation.column_corpus_transformation import (
     ColumnCorpusTransformation,
 )
 
 
-class HuggingFaceSequenceTaggingPipeline(
+class HuggingFaceSequenceLabelingPipeline(
     StandardPipeline[str, datasets.DatasetDict, Corpus, Dict[str, np.ndarray], Dict[str, Any]]
 ):
     def __init__(
@@ -33,7 +33,7 @@ class HuggingFaceSequenceTaggingPipeline(
         data_loader = HuggingFaceDataLoader()
         transformation = ColumnCorpusTransformation(input_column_name, target_column_name)
         embedding = FlairTransformerWordEmbedding(embedding_name)
-        task = SequenceTagging(output_path, hidden_size=hidden_size)
+        task = SequenceLabeling(output_path, hidden_size=hidden_size)
         model = FlairModel(embedding, task)
-        evaluator = SequenceTaggingEvaluator()
+        evaluator = SequenceLabelingEvaluator()
         super().__init__(dataset, data_loader, transformation, model, evaluator)
